requirements.txt
matplotlib
numpy
pandas
brainflow
scipy
pyserial
PyQt6
pyqtgraph